value_list: Migrate `parse_values` out of `common`.
