+++ src/utils/common/common_test.c
@@ -272,56 +272,6 @@ DEF_TEST(strunescape) {
   return 0;
 }
 
-DEF_TEST(parse_values) {
-  struct {
-    char buffer[64];
-    int status;
-    gauge_t value;
-  } cases[] = {
-      {"1435044576:42", 0, 42.0}, {"1435044576:42:23", -1, NAN},
-      {"1435044576:U", 0, NAN},   {"N:12.3", 0, 12.3},
-      {"N:42.0:23", -1, NAN},     {"N:U", 0, NAN},
-      {"T:42.0", -1, NAN},
-  };
-
-  for (size_t i = 0; i < STATIC_ARRAY_SIZE(cases); i++) {
-    data_source_t dsrc = {
-        .name = "value",
-        .type = DS_TYPE_GAUGE,
-        .min = 0.0,
-        .max = NAN,
-    };
-    data_set_t ds = {
-        .type = "example",
-        .ds_num = 1,
-        .ds = &dsrc,
-    };
-
-    value_t v = {
-        .gauge = NAN,
-    };
-    value_list_t vl = {
-        .values = &v,
-        .values_len = 1,
-        .time = 0,
-        .interval = 0,
-        .host = "example.com",
-        .plugin = "common_test",
-        .type = "example",
-        .meta = NULL,
-    };
-
-    int status = parse_values(cases[i].buffer, &vl, &ds);
-    EXPECT_EQ_INT(cases[i].status, status);
-    if (status != 0)
-      continue;
-
-    EXPECT_EQ_DOUBLE(cases[i].value, vl.values[0].gauge);
-  }
-
-  return 0;
-}
-
 DEF_TEST(value_to_rate) {
   struct {
     time_t t0;
@@ -419,7 +369,6 @@ int main(void) {
   RUN_TEST(escape_slashes);
   RUN_TEST(escape_string);
   RUN_TEST(strunescape);
-  RUN_TEST(parse_values);
   RUN_TEST(value_to_rate);
   RUN_TEST(format_values);
